partially retype FOV range as short rather than int

+++ RogueSurvivor/Engine/Items/ItemLight.cs
@@ -26,7 +26,7 @@ namespace djack.RogueSurvivor.Engine.Items
     }
 
     public int MaxBatteries { get { return Model.MaxBatteries; } }
-    public int FovBonus { get { return Model.FovBonus; } }
+    public short FovBonus { get { return Model.FovBonus; } }
 
     public override string ImageID {
       get {
